Typecast image_kwargs['protected'] to boolean in _upload_image_put_v2
When attempting to use create_image with protected=True as a kwargs argument,
I got:

```
Failed validating 'type' in schema['properties']['protected']:
    {'description': 'If true, image will not be deletable.',
     'type': 'boolean'}

On instance['protected']:
    'True'
```

Not sure why but all the image_kwargs are getting typecast to str and then
min_disk and min_ram are getting typecast back to int. It appears that the
typecast to str is not necessary as removing it doesn't appear to cause
failures. This should also fix tags which is required to be a list.